ruta de usuarios

+++ src/controllers/controllers.producto.js
@@ -1,22 +1,28 @@
-// controlador de los productos
+// controlador de productos
 const getPool = require('../connection');
 const { closeConnection } = require('../funciones');
 
 
-// mostrar todos los productos
-async function getsProducts (req,res) {
+// mostrar los productos
+async function getsProducts(req, res) {
     try {
+        // validación de token
         const token = req.cookies.token;
         if (!token) {
             return res.json(
                 {
+                    status:400,
                     error: true,
                     errorMessage: 'No hay token, acceso no autorizado'
                 }
             );
         }
+        // llamamos la conexion
         const pool = await getPool();
-        const result = await pool.query('SELECT * FROM t_productos');
+        const sqlQuery = `SELECT * FROM t_productos`;
+        const result = await pool.query(sqlQuery);
+
+        // mandamos la repuesta
         res.json({
             status: 200,
             error: false,
@@ -25,8 +31,11 @@ async function getsProducts (req,res) {
             token: req.results,
             results: result.rows
         });
+
+        // Cerramos la conexion
         closeConnection(pool, res);
     } catch (error) {
+        // manejo de errores
         console.error('Error al ejecutar la consulta:', error);
         res.json({
             status: 500,
@@ -37,24 +46,53 @@ async function getsProducts (req,res) {
     }
 }
 
-// añadir productos
-async function addProducts (req,res) {
+// agregar productos
+async function addProducts(req, res) {
+
     try {
+        // validación de token
         const token = req.cookies.token;
-
         if (!token) {
-            return res.json({
-                error: true,
-                errorMessage: 'No hay token, acceso no autorizado'
-            });
+            return res.json(
+                {
+                    status:400,
+                    error: true,
+                    errorMessage: 'No hay token, acceso no autorizado'
+                }
+            );
         }
-        const { codProducto, nombre, descripcion, precio } = req.body;
-        const pool = await getPool();
-        const resultSeq = await pool.query('SELECT max(NCODIGO) + 1 as seq FROM t_productos');
-        const seq = resultSeq.rows[0].seq;  
-        const query = 'INSERT INTO t_productos (NCODIGO, NCODIGO_PRODUCTO, VNOMBRE, VDESCRIPCION, NPRECIO) VALUES ($1, $2, $3, $4, $5)';
-        const result = await pool.query(query, [seq, codProducto, nombre, descripcion, precio]);
 
+        // validamos que le usuario si posee los privilegios adecuados
+        if (Number(req.results.privilegio) == 4 || Number(req.results.privilegio) == 2) {
+            return res.json({
+                status: 400,
+                error: true,
+                errorMessage: 'No tiene acceso a esta acción por motivos de privilegios'
+            })
+        }
+        // capturamos los campo del mandado por el body
+        const { codProducto, nombre, descripcion, precio } = req.body;
+        // llamamos la conexion
+        const pool = await getPool();
+        // obetenemos la secuencia
+        const resultSeq = await pool.query('SELECT max(NCODIGO) + 1 as seq FROM t_productos');
+        // alamcenamos la secunecia en una variable
+        const seq = resultSeq.rows[0].seq;
+        // construimos el query 
+        const sqlQuery = `
+        INSERT INTO t_productos 
+        (
+            NCODIGO
+            ,NCODIGO_PRODUCTO
+            ,VNOMBRE
+            ,VDESCRIPCION
+            ,NPRECIO
+        ) VALUES ($1, $2, $3, $4, $5)
+        `;
+        // ejecutamos el query
+        const result = await pool.query(sqlQuery, [seq, codProducto, nombre, descripcion, precio]);
+
+        // validamos el insert
         if (result.rowCount > 0) {
             res.json({
                 status: 200,
@@ -63,42 +101,69 @@ async function addProducts (req,res) {
                 results: result.rows[0]
             });
         } else {
+            // si no se hizo la inserción
             res.json({
                 status: 500,
                 error: true,
-                message: 'Error al guardar los campos',
-                results: null
+                errorMessage: 'Error al guardar el producto'
             });
         }
 
+        // Cerramos la conexion
         closeConnection(pool, res);
 
     } catch (error) {
+        // manejos de errores
         console.error('Error al ejecutar el insert:', error);
         res.json({
             status: 500,
             error: true,
-            errorDes: 'Error interno del servidor',
+            message: 'Error interno del servidor',
             errorMessage: error.message
         });
     }
 }
 
 // actualizar productos
-async function updateProducts (req,res){
-    try{
+async function updateProducts(req, res) {
+    try {
+        // validación de token
         const token = req.cookies.token;
-        if(!token){
-            return res.json({
-                error: true,
-                errorMessage: 'No hay token, acceso no autorizado'
-            });
+        if (!token) {
+            return res.json(
+                {
+                    status:400,
+                    error: true,
+                    errorMessage: 'No hay token, acceso no autorizado'
+                }
+            );
         }
-        const { codigo ,codProducto, nombre, descripcion, precio } = req.body;
-        const pool = await getPool();
-        const query = 'UPDATE t_productos SET NCODIGO_PRODUCTO = $2, VNOMBRE = $3, VDESCRIPCION = $4, NPRECIO = $5 WHERE NCODIGO = $1';
-        const result = await pool.query(query,[codigo,codProducto,nombre,descripcion,precio]);
 
+        // validamos que le usuario si posee los privilegios adecuados
+        if (Number(req.results.privilegio) == 4 || Number(req.results.privilegio) == 2) {
+            return res.json({
+                status: 400,
+                error: true,
+                errorMessage: 'No tiene acceso a esta acción por motivos de privilegios'
+            })
+        }
+
+        // capturamos los campo del body
+        const { codigo, codProducto, nombre, descripcion, precio } = req.body;
+        // llamamos la conexion
+        const pool = await getPool();
+        const sqlQuery = `
+        UPDATE t_productos SET 
+            NCODIGO_PRODUCTO = $2
+            ,VNOMBRE = $3
+            ,VDESCRIPCION = $4
+            ,NPRECIO = $5
+        WHERE NCODIGO = $1
+        `;
+        // ejecutamos el query
+        const result = await pool.query(sqlQuery, [codigo, codProducto, nombre, descripcion, precio]);
+
+        // validamos la actualización
         if (result.rowCount > 0) {
             res.json({
                 status: 200,
@@ -107,17 +172,19 @@ async function updateProducts (req,res){
                 results: result.rows[0]
             });
         } else {
+            // si hay un error al actulizar
             res.json({
                 status: 500,
                 error: true,
-                message: 'Error al actualizar los campos',
-                results: null
+                errorMessage: 'Error al actualizar los campos'
             });
         }
 
+        // Cerramos la conexion
         closeConnection(pool, res);
 
-    }catch(error){
+    } catch (error) {
+        // manejos de errores
         console.error('Error al ejecutar el update:', error);
         res.json({
             status: 500,
@@ -129,14 +196,38 @@ async function updateProducts (req,res){
 }
 
 // eliminar producto
-async function deleteProduct(req,res){
+async function deleteProduct(req, res) {
     try {
+        // validación de token
+        const token = req.cookies.token;
+        if (!token) {
+            return res.json(
+                {
+                    status:400,
+                    error: true,
+                    errorMessage: 'No hay token, acceso no autorizado'
+                }
+            );
+        }
+
+        // validamos que le usuario si posee los privilegios adecuados
+        if (Number(req.results.privilegio) == 3 || Number(req.results.privilegio) == 2) {
+            return res.json({
+                status: 400,
+                error: true,
+                errorMessage: 'No tiene acceso a esta acción por motivos de privilegios'
+            })
+        }
+
+        // capturamos el parametro
         const id = req.params.id;
+        // llamamos la conexion
         const pool = await getPool();
-        const result = await pool.query('DELETE FROM t_productos WHERE NCODIGO = $1',[id]);
+        // ejecutamos el query
+        const result = await pool.query('DELETE FROM t_productos WHERE NCODIGO = $1', [id]);
 
+        // validamos la eliminación
         if (result.rowCount > 0) {
-
             res.json({
                 status: 200,
                 error: false,
@@ -144,42 +235,51 @@ async function deleteProduct(req,res){
                 results: result.rows[0]
             });
         } else {
+            // manejo de errores
             res.json({
                 status: 404,
                 error: true,
-                message: 'No se encontró el producto con el ID proporcionado',
-                result:null
+                errorMessage: 'No se encontró el producto con el ID proporcionado'
             });
         }
 
+        // Cerrar la conexion
         closeConnection(pool, res);
-        
+
     } catch (error) {
+        // manejo de errores
         console.error('Error al ejecutar el delete:', error);
         res.json({
             status: 500,
             error: true,
             errorDes: 'Error interno del servidor',
             errorMessage: error.message
-        }); 
+        });
     }
 }
 
 // mostra un producto especifico
-async function getProduct (req,res){
+async function getProduct(req, res) {
     try {
+        // validación de token
         const token = req.cookies.token;
         if (!token) {
             return res.json(
                 {
+                    status:400,
                     error: true,
                     errorMessage: 'No hay token, acceso no autorizado'
                 }
             );
         }
+
+        // capturamos el parametro
         const id = req.params.id;
+        // llamos la conexion
         const pool = await getPool();
-        const result = await pool.query('SELECT * FROM t_productos where NCODIGO = $1',[id]);
+        // ejecutamos el query 
+        const result = await pool.query('SELECT * FROM t_productos where NCODIGO = $1', [id]);
+        // mandamos la respuesta
         res.json({
             status: 200,
             error: false,
@@ -188,8 +288,10 @@ async function getProduct (req,res){
             token: req.results,
             results: result.rows
         });
+        // Cerramos la conexion
         closeConnection(pool, res);
     } catch (error) {
+        // manejos de errores
         console.error('Error al ejecutar la consulta:', error);
         res.json({
             status: 500,
@@ -200,6 +302,7 @@ async function getProduct (req,res){
     }
 }
 
+// exportamos las funciones  
 module.exports = {
     getsProducts,
     addProducts,
